bindinfo: migrate test-infra to testify(staging-4)
